libs/autofile: Move into internal
